Extract helper utilities into separate module
